Update usage and tests
